Screen rendering refactored to copy a surface to the screen, rather than having SDL draw each triangle
