+++ Cargo.toml
@@ -7,6 +7,6 @@ edition = "2021"
 
 [dependencies]
 
-sdl3 = { version = "0.11.14" }
+sdl3 = { version = "0.15.1" }
 glam = { version = "*" }
 rand = "*"
