+++ src/clippingtest.py
