+++ src/main.rs
@@ -1,49 +1,57 @@
 #![allow(unused_must_use)]
 #![allow(dead_code)]
+#![allow(non_snake_case)]
 
 extern crate sdl3;
 extern crate glam;
 extern crate rand;
 
-mod engine;
+mod geo_engine;
+mod gfx_engine;
 
-pub use engine::*;
+pub use geo_engine::*;
 
-use sdl3::pixels::Color;
-use sdl3::render::FPoint;
+use sdl3::pixels::{Color, PixelFormat};
+use sdl3::render::{FPoint, Texture, TextureAccess};
 use sdl3::keyboard::Keycode;
 
 use glam::Vec3;
 
+
 fn main() {
-    static SCREEN_WIDTH: u32 = 1280;
-    static SCREEN_HEIGHT: u32 = 720;
+    static SCREEN_WIDTH: u16 = 1280;
+    static SCREEN_HEIGHT: u16 = 720;
     static FRUSTUM_INSET: f32 = 5.;
 
     let move_speed = 1.;
     
     //Read the OBJ file's data
     let uwb = std::env::current_dir().unwrap();
-    let p = uwb.to_str().unwrap();
-    let full_path = format!("{}{}", p, "\\resources\\Leviathan");
-    let full_path_str = full_path.as_str();
+    let path_prefix = uwb.to_str().unwrap();
+    let local_path = "/resources/Leviathan";
     let mut verts = Vec::new();
     let mut tris = Vec::new();
-    let mut mats = Vec::new();
-    read_geometry(full_path_str, &mut verts, &mut tris, &mut mats);
-
-    let mut viewer = Camera::new(Vec3::new(0., 0., 0.), 2.5, SCREEN_WIDTH, SCREEN_HEIGHT, 1000., FRUSTUM_INSET/* clipdist */, FRUSTUM_INSET);
+    let mut matIdcs = Vec::new();
+    let mut materials = Vec::new();
+    read_geometry(path_prefix, local_path, &mut verts, &mut tris, &mut matIdcs, &mut materials);
 
     let sdl = sdl3::init().unwrap();
     let vsub = sdl.video().unwrap();
-    let window = vsub.window("Game", SCREEN_WIDTH, SCREEN_HEIGHT).resizable().build().unwrap();
+    let window = vsub.window("Raster Renderer", SCREEN_WIDTH as u32, SCREEN_HEIGHT as u32).resizable().build().unwrap();
     
-    sdl.mouse().set_relative_mouse_mode(&window, true);
     sdl.mouse();
+    sdl.mouse().set_relative_mouse_mode(&window, true);
+    sdl.mouse().warp_mouse_in_window(&window, (SCREEN_WIDTH/2) as f32, (SCREEN_HEIGHT/2) as f32);
 
     let mut canvas = window.into_canvas();
-    let mut event_pump = sdl.event_pump().unwrap();
+
+    let texture_creator = canvas.texture_creator();
+    let pixel_format = canvas.default_pixel_format();
+    let mut render_surf = gfx_engine::Surface::new(SCREEN_WIDTH, SCREEN_HEIGHT, pixel_format, &texture_creator);
+    let mut viewer = Camera::new(Vec3::new(0., 0., 0.), 2.5, SCREEN_WIDTH as u32, SCREEN_HEIGHT as u32, 1000., FRUSTUM_INSET/* clipdist */, FRUSTUM_INSET);
     
+    let mut event_pump = sdl.event_pump().unwrap();
+
     //let mut now: std::time::Instant;
     'gl: loop {
         for event in event_pump.poll_iter() {match event {
@@ -61,21 +69,15 @@ fn main() {
         //let mousex; let mousey;
         //sdl3-sys::SDL_GetMouseState(&mut mousex, &mut mousey);
         
-        canvas.set_draw_color(Color::BLACK);
-        canvas.clear();
-        
+        // Project geometry to the screen
         let mut screen_tris = Vec::new();
         for i in 0..tris.len() {
             match viewer.clip_tri_to_frustum(tris[i], &verts) {
                 Some(clipped_tri) => { // If the triangle is valid, render it
                     for tri in clipped_tri.iter() {
                         let (screen_tri, depth) = viewer.project_tri(*tri);
-
-                        if (screen_tri[0].x >= 0. || screen_tri[1].x >= 0. || screen_tri[2].x >= 0.) // If any of the projected vertices are actually on-screen (should be redundant
-                        && (screen_tri[0].x <= SCREEN_WIDTH as f32 || screen_tri[1].x <= SCREEN_WIDTH as f32 || screen_tri[2].x <= SCREEN_WIDTH as f32) // with frustum culling)
-                        && (screen_tri[0].y >= 0. || screen_tri[1].y >= 0. || screen_tri[2].y >= 0.)
-                        && (screen_tri[0].y <= SCREEN_HEIGHT as f32 || screen_tri[1].y <= SCREEN_HEIGHT as f32 || screen_tri[2].y <= SCREEN_HEIGHT as f32) {
-                            screen_tris.push((screen_tri, depth, mats[i].color));
+                        if render_surf.clip_tri_to_screen(screen_tri) {
+                            screen_tris.push((screen_tri, depth, matIdcs[i]));
                         }
                     }
                 },
@@ -83,18 +85,12 @@ fn main() {
             }
         }
 
+        render_surf.render_tris(&mut screen_tris, &materials);
 
-        let l = screen_tris.len();
-        screen_tris.sort_by(|a, b| a.1.partial_cmp(&b.1).unwrap());
-        for i in 0..l {
-            if canvas.draw_color() != screen_tris[(l-1)-i].2 { canvas.set_draw_color(screen_tris[l-i-1].2) }
-            //println!("Draw color set, rasterizing triangle...");
-            let frag: Vec<FPoint> = get_points_in_triangle(screen_tris[l-i-1].0);
-            //println!("Pixels determined, drawing points...");
-            canvas.draw_points(frag.as_slice());
-        }
-
-    //PRESENT and FPS TIMESTEP
+    // Present the scene and FPS Timestep
+        canvas.set_draw_color(Color::BLACK);
+        canvas.clear();
+        canvas.copy(&render_surf.render_tex, None, None);
         canvas.present();
         std::thread::sleep(std::time::Duration::new(0, 1_000_000_000u32 / 144 as u32));
     }
